chore: rename folder api-model to interfaces
