+++ ts/apps/tournament/src/app/controllers/participant-controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 // import { TournamentRepository } from '../repository/tournament-repository';
-import { Participant, tournamentRepository } from '../api/api-model';
+import { Participant, tournamentRepository } from '../interfaces';
 
 // const tournamentRepository = new TournamentRepository();
