+++ ts/apps/tournament/src/app/interfaces/index.ts
